# ARROW-9268: [C++] add string_is{alpnum,alpha...,upper} kernels