--- a/cpp/src/arrow/compute/kernels/scalar_string.cc
+++ b/cpp/src/arrow/compute/kernels/scalar_string.cc
@@ -727,7 +727,7 @@ struct IsNumericUnicode
 template <typename ArrowType>
 struct IsAscii : BinaryToBoolean<ArrowType, IsAscii<ArrowType>> {
   using offset_type =
-      typename BinaryToBoolean<StringType, IsAscii<StringType>>::offset_type;
+      typename BinaryToBoolean<ArrowType, IsAscii<ArrowType>>::offset_type;
   static bool Predicate(KernelContext* ctx, const uint8_t* input,
                         offset_type input_string_nascii_characters) {
     return std::all_of(input, input + input_string_nascii_characters,
@@ -799,7 +799,7 @@ struct IsSpaceAscii : CharacterPredicateAscii<ArrowType, IsSpaceAscii<ArrowType>
 template <typename ArrowType>
 struct IsTitleUnicode : BinaryToBoolean<ArrowType, IsTitleUnicode<ArrowType>> {
   using offset_type =
-      typename BinaryToBoolean<ArrowType, IsTitleUnicode<StringType>>::offset_type;
+      typename BinaryToBoolean<ArrowType, IsTitleUnicode<ArrowType>>::offset_type;
   static bool Predicate(KernelContext* ctx, const uint8_t* input,
                         offset_type input_string_ncodeunits) {
     // rules: