--- a/cpp/src/arrow/compute/kernels/scalar_string_benchmark.cc
+++ b/cpp/src/arrow/compute/kernels/scalar_string_benchmark.cc
@@ -57,8 +57,8 @@ static void AsciiUpper(benchmark::State& state) {
   UnaryStringBenchmark(state, "ascii_upper");
 }
 
-static void IsAlphaAscii(benchmark::State& state) {
-  UnaryStringBenchmark(state, "ascii_isalpha");
+static void IsAlphaNumericAscii(benchmark::State& state) {
+  UnaryStringBenchmark(state, "ascii_isalnum");
 }
 
 static void BinaryContainsExact(benchmark::State& state) {
@@ -75,19 +75,19 @@ static void Utf8Lower(benchmark::State& state) {
   UnaryStringBenchmark(state, "utf8_lower");
 }
 
-static void IsAlphaUnicode(benchmark::State& state) {
-  UnaryStringBenchmark(state, "utf8_isalpha");
+static void IsAlphaNumericUnicode(benchmark::State& state) {
+  UnaryStringBenchmark(state, "utf8_isalnum");
 }
 #endif
 
 BENCHMARK(AsciiLower);
 BENCHMARK(AsciiUpper);
-BENCHMARK(IsAlphaAscii);
+BENCHMARK(IsAlphaNumericAscii);
 BENCHMARK(BinaryContainsExact);
 #ifdef ARROW_WITH_UTF8PROC
 BENCHMARK(Utf8Lower);
 BENCHMARK(Utf8Upper);
-BENCHMARK(IsAlphaUnicode);
+BENCHMARK(IsAlphaNumericUnicode);
 #endif
 
 }  // namespace compute