--- a/cpp/src/arrow/compute/kernels/scalar_string_benchmark.cc
+++ b/cpp/src/arrow/compute/kernels/scalar_string_benchmark.cc
@@ -57,6 +57,10 @@ static void AsciiUpper(benchmark::State& state) {
   UnaryStringBenchmark(state, "ascii_upper");
 }
 
+static void IsAlphaAscii(benchmark::State& state) {
+  UnaryStringBenchmark(state, "string_isalpha_ascii");
+}
+
 static void BinaryContainsExact(benchmark::State& state) {
   BinaryContainsExactOptions options("abac");
   UnaryStringBenchmark(state, "binary_contains_exact", &options);
@@ -70,14 +74,20 @@ static void Utf8Upper(benchmark::State& state) {
 static void Utf8Lower(benchmark::State& state) {
   UnaryStringBenchmark(state, "utf8_lower");
 }
+
+static void IsAlphaUnicode(benchmark::State& state) {
+  UnaryStringBenchmark(state, "string_isalpha_unicode");
+}
 #endif
 
 BENCHMARK(AsciiLower);
 BENCHMARK(AsciiUpper);
+BENCHMARK(IsAlphaAscii);
 BENCHMARK(BinaryContainsExact);
 #ifdef ARROW_WITH_UTF8PROC
 BENCHMARK(Utf8Lower);
 BENCHMARK(Utf8Upper);
+BENCHMARK(IsAlphaUnicode);
 #endif
 
 }  // namespace compute