… cpp/src/arrow/compute/kernels/scalar_string_test.cc
… cpp/src/arrow/util/utf8.h
… python/pyarrow/compute.py
… python/pyarrow/tests/test_compute.py