@wesm 2020-07-07
I haven't looked at the details yet, but could we use a normalized naming convention for the functions? 

* ascii_$FUNC, for functions that expect ASCII input
* utf8_$FUNC, for functions that expect UTF8 input
* binary_$FUNC, for functions that work on any binary data