@maartenbreddels 2020-07-07
I wanted to focus on the other issues first, but I propose a change in naming by this:
 `<arrow type>_<func>_<variant>`, since all you care about is that it's a function that operates on a string and all the ascii variants with with all utf8 strings. Also, I don't know if in the future there will be utf16/utf32 support, in that case, this naming scheme makes more sense. I'm ok changing it back or open a PR to make this consistent (`string_lower_ascii`, `string_lower_uncode`, ... etc).