@wesm 2020-07-07
OK, I would also like to solicit @xhochy's and @pitrou's preferences in the function naming. I don't want to spend too much time on it, but as long as we have a consistent naming scheme that accommodates the anticipated use cases, that sounds fine to me.