@wesm 2020-07-07
I think that having e.g. `string_lower_utf8` and `string_lower_ascii` is fine, too.