@pitrou 2020-07-07
We should be consistent. We already have "utf8_lower" and "utf8_upper".