@maartenbreddels 2020-07-07
Are there plans to support utf16/32? Seeing the code as it is now, it would be trivial to add.