@xhochy 2020-07-07
> Are there plans to support utf16/32? Seeing the code as it is now, it would be trivial to add.

This would be a longer discussion. Personally I would vote "no" to limit the scope / dimensionality of Arrow's memory model for now.