@maartenbreddels 2020-07-07
If it is not off the table, I propose at least utf8->unicode renaming, to be future compatible.