@pitrou 2020-07-07
"utf8" is used everywhere, please keep it like this.