@pitrou 2020-07-07
> Are there plans to support utf16/32?

I think that should be out of scope. There's no Arrow data type for utf16 and utf32.