@maartenbreddels 2020-07-07
I am not talking about the C++ code, only the kernel name. UTF8 is just the encoding, Unicode refers to the semantics of the kernel. But if utf16/32 if excluded, I guess we can keep the old naming scheme.