@pitrou 2020-07-07
I am talking about the kernel name as well. "unicode" is non-descriptive.