@stevengj 2020-07-08
[U+08BE](https://www.fileformat.info/info/unicode/char/08be/index.htm) was defined in Unicode 13, and category Lo is correct.