@pitrou 2020-07-08
As for the performance, I think 400 MB/s is perfectly acceptable. I don't know in which kind of situation these functions would be performance-critical.