@pitrou 2020-07-08
The Python `unicodedata` module gives you the Unicode spec version:
```
Python 3.7.5 (default, Nov  7 2019, 10:50:52) 
[GCC 8.3.0] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> import unicodedata
>>> unicodedata.unidata_version
'11.0.0'
```
```
Python 3.8.0+ (heads/3.8:fe934e1d03, Nov  2 2019, 16:21:38) 
[GCC 7.4.0] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> import unicodedata
>>> unicodedata.unidata_version
'12.1.0'
```