@pitrou 2020-07-08
Also, for the record, the Python `str` methods are implemented in `Objects/unicodeobject.c`, for example:
https://github.com/python/cpython/blob/master/Objects/unicodeobject.c#L12051-L12094