@maartenbreddels 2020-07-08
Thanks, I didn't know about unicodedata. Yeah, I've taken a look at that, because sometimes the code is more clear than the documentation.