@pitrou 2020-07-08
Also, the predicates are defined here:
https://github.com/python/cpython/blob/master/Objects/unicodectype.c