@maartenbreddels 2020-07-08
The benchmark ran `isalpha`, which was a bad idea (high probability for a false on the first character), `isalnum` is a better test case, with characters in the range `A`-'z' (including some non-alpnum characters).
Before:
```
IsAlphaNumericUnicode_median   68083381 ns     68066554 ns            3 bytes_per_second=232.786M/s items_per_second=15.4052M/s
```
After:
```
IsAlphaNumericUnicode_median   44940488 ns     44931459 ns            3 bytes_per_second=352.647M/s items_per_second=23.3372M/s
```

This is still 3x slower than what I have in Vaex, so I guess the compiler doesn't like my code.