@wesm 2020-07-10
I added this issue to the 1.0.0 milestone so we can try to get this merged in the next few days