@pitrou 2020-07-10
> Does that build maybe pick up an older version of utf8proc somehow?

I don't think we care about that. You should simply test less edgy cases in the unit tests (but make sure you test characters above 0xFFFF).