@xhochy 2020-07-10
> Does that build maybe pick up an older version of utf8proc somehow?

Yes, it is picking up `utf8proc` 2.1.0.