@maartenbreddels 2020-07-10
I hope the rstudio failures are unrelated. The compiler error for https://github.com/apache/arrow/pull/7656/checks?check_run_id=858028868 I have trouble decoding.