@wesm 2020-07-10
Seems like a legitimate error with that particular MinGW toolchain (not sure why it doesn't show up in the other MinGW builds), can take a look today