@wesm 2020-07-10
I googled a little bit about the compilation errors and it seems to be a gcc bug. I think that some of these kernels are going to have to be rewritten to use inline lambdas instead of CRTP. There should be no performance difference. I'll take a closer look at it this weekend