@maartenbreddels 2020-07-11
Indeed, you're right. It seems to hit https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65333 but that reports GCC 5, and is quite old, looking at the logs it seems to be using gcc 8.3 (I can't find when it's fixed, or is it a different bug or a regression).